requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pillow==10.1.0
matplotlib==3.8.2
numpy==1.26.2 